Changed [DELETE] api/review/:id - add error handlers

+++ src/review/review.service.ts
@@ -1,8 +1,12 @@
 import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { Model } from 'mongoose';
-
-import { PRODUCT_NOT_FOUND } from 'src/constants';
+import {
+  ACCESS_DENIED,
+  BAD_DELETE_REVIEW_REQUEST,
+  PRODUCT_NOT_FOUND,
+  REVIEW_NOT_FOUND,
+} from 'src/constants';
 import { ReviewModel } from './review.model';
 import { CreateReviewDto } from './dto/createReviewDto';
 import { ProductService } from 'src/product/product.service';
@@ -32,7 +36,30 @@ export class ReviewService {
     });
   }
 
-  async delete(reviewId: string) {
-    return this.reviewModel.findByIdAndDelete(reviewId);
+  async delete(userId: string, productId: string, reviewId: string) {
+    const review = await this.reviewModel.findById(reviewId);
+
+    if (!review) {
+      throw new HttpException(REVIEW_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
+    if (review.author.toString() !== userId) {
+      throw new HttpException(ACCESS_DENIED, HttpStatus.FORBIDDEN);
+    }
+
+    const product = await this.productService.findById(productId);
+
+    if (!product) {
+      throw new HttpException(PRODUCT_NOT_FOUND, HttpStatus.NOT_FOUND);
+    }
+
+    if (review.productId.toString() !== productId) {
+      throw new HttpException(
+        BAD_DELETE_REVIEW_REQUEST,
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    await this.reviewModel.findByIdAndDelete(reviewId);
   }
 }
